# 024/Project/src/docs/swagger.yaml
basePath: /api/v0
definitions:
  baseresponse.Response:
    properties:
      error:
        type: string
      result: {}
      success:
        type: boolean
    type: object
...
  bindparam.LinkMovie2CastRequest:
    properties:
      cast-id:
        type: integer
      movie-id:
        type: integer
    required:
    - cast-id
    - movie-id
    type: object
...
host: localhost:8097
info:
  contact:
    email: m9652973@gmail.com
    name: Matin Hasanali Baki ( Matin HAB )
    url: https://No-Where-Be-Allah-Ghasam-2
  description: REST API for managing movies and casts
  license:
    name: Apache 9797.0
    url: https://No-Where-Be-Allah-Ghasam-3
  termsOfService: https://No-Where-Be-Allah-Ghasam-1
  title: Movie App API
  version: "0.0"
paths:
...
  /movie-cast/:
    get:
      description: Returns the list of all movie and cast relationships
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Get all movie-cast relations
      tags:
      - MovieCast
    post:
      consumes:
      - application/json
      description: Create relation between a movie and a cast
      parameters:
      - description: Movie Cast Link
        in: body
        name: data
        required: true
        schema:
          $ref: '#/definitions/bindparam.LinkMovie2CastRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Link movie to cast
      tags:
      - MovieCast
  /movie-cast/{id}:
    delete:
      description: Deletes a movie-cast relationship using its ID
      parameters:
      - description: MovieCast ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Delete movie-cast relation by ID
      tags:
      - MovieCast
    get:
      description: Returns a specific movie-cast relationship using its ID
      parameters:
      - description: MovieCast ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Get movie-cast relation by ID
      tags:
      - MovieCast
  /movie-cast/movie/{movie_id}/cast/{cast_id}:
    delete:
      description: Deletes the relationship between a specific movie and cast member
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: integer
      - description: Cast ID
        in: path
        name: cast_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Delete movie-cast relation by movie ID and cast ID
      tags:
      - MovieCast
    get:
      description: Returns the relationship between a specific movie and cast member
      parameters:
      - description: Movie ID
        in: path
        name: movie_id
        required: true
        type: integer
      - description: Cast ID
        in: path
        name: cast_id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/baseresponse.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/baseresponse.Response'
      summary: Get movie-cast relation by movie ID and cast ID
      tags:
      - MovieCast
...
swagger: "2.0"